refactor(designspec.js): define three-/two-column layout parameters in designSpecs.js
Previously we defined them in breakpoints.js. But they are the layout parameters, not about media
query per se.
